requirements.txt
google-genai==1.8.0
loadotenv==1.0.1
pandas==2.2
numpy
python-dotenv==1.1.0
pyarrow==19.0.1